move shaders out of file
